+++ .ts-for-gir.gtk4.rc.js
@@ -1,6 +1,6 @@
 export default {
     modules: ['Gtk-4.0'],
     girDirectories: ['./vala-girs/gir-1.0'],
-    pretty: true,
+    pretty: false, // FIXME
     ignore: [],
 }