+++ templates/Gjs/cast.ts
@@ -1,5 +1,5 @@
 
-<% if(exportNamespace){ %>
+<% if(useNamespace){ %>
     import GObject from './GObject-2.0';
   <% } else { %>  
     import * as GObject from './GObject-2.0';